instancewise square selection IMP

+++ self-supervised-depth-completion-master2_working/helper.py
@@ -222,10 +222,10 @@ def adjust_learning_rate(lr_init, optimizer, epoch):
     return lr
 
 
-def save_checkpoint(state, is_best, epoch, output_directory, type_feature, i=0, every=50):
+def save_checkpoint(state, is_best, epoch, output_directory, type_feature, i=0, every=50, qnet=None):
 
     if is_best:
-        checkpoint_filename = os.path.join(output_directory, f'checkpoint-{str(epoch - 1)}_i_{i}_typefeature_{type_feature}.pth.tar')
+        checkpoint_filename = os.path.join(output_directory, f'checkpoint_qnet-{str(epoch - 1)}_i_{i}_typefeature_{type_feature}.pth.tar')
         torch.save(state, checkpoint_filename)
 
         # best_filename = os.path.join(output_directory, f'model_best-{str(epoch - 1)}_i_{i}_typefeature_{type_feature}.pth.tar')
@@ -233,8 +233,8 @@ def save_checkpoint(state, is_best, epoch, output_directory, type_feature, i=0, 
     #if epoch > 0:
     if 1:
         prev_checkpoint_filename = os.path.join(
-            output_directory, f'checkpoint-{str(epoch - 1)}_i_{i-every}_typefeature_{type_feature}.pth.tar')
-        if os.path.exists(prev_checkpoint_filename):
+            output_directory, f'checkpoint_qnet-{str(epoch - 1)}_i_{i-every}_typefeature_{type_feature}.pth.tar')
+        if os.path.exists(prev_checkpoint_filename) and ((i-every) % 500 != 0):
             os.remove(prev_checkpoint_filename)
